mcp-server/requirements.txt
fastmcp>=2.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
msal>=1.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
